feat: enhance home feature with remote data source, repository, and use cases; update models and integrate BLoC for state management; adding get it and env load on main; adding DI home module;

+++ lib/features/home/data/models/balance_model.dart
@@ -2,7 +2,7 @@ import 'package:mobile_banking_apps/features/home/domain/entities/balance_entity
 
 class BalanceModel {
   final String username;
-  final BigInt balance;
+  final double balance;
 
   BalanceModel({
     required this.username,
