+++ lib/features/home/domain/entities/balance_entity.dart
@@ -1,6 +1,6 @@
 class BalanceEntity {
   final String username;
-  final BigInt balance;
+  final double balance;
 
   BalanceEntity({
     required this.username,
